Standalone tools install

+++ debian/appscale_install_functions.sh
@@ -6,25 +6,23 @@
 #
 # Written by Yoshi <nomura@pobox.com>
 
-if [ -z "$APPSCALE_HOME_RUNTIME" ]; then
-    export APPSCALE_HOME_RUNTIME=/root/appscale
-fi
+set -e
 
-if [ -z "$APPSCALE_HOME" ]; then
-    export APPSCALE_HOME=/root/appscale
+if [ -z "$APPSCALE_TOOLS_HOME" ]; then
+    export APPSCALE_TOOLS_HOME=/root/appscale
 fi
-
 
 installexpect()
 {
-  mkdir -pv ${APPSCALE_HOME}/downloads
-  cd ${APPSCALE_HOME}/downloads
-  curl -o expect5.45.tar.gz http://appscale.cs.ucsb.edu/appscale_files/expect5.45.tar.gz || exit 1
-  tar zxvf expect5.45.tar.gz || exit 1
+  echo "Installing expect"
+  mkdir -pv ${APPSCALE_TOOLS_HOME}/downloads
+  cd ${APPSCALE_TOOLS_HOME}/downloads
+  curl -o expect5.45.tar.gz http://appscale.cs.ucsb.edu/appscale_files/expect5.45.tar.gz
+  tar zxvf expect5.45.tar.gz
   pushd expect5.45
-  ./configure || exit 1
-  make || exit 1
-  make install  || exit 1
+  ./configure
+  make
+  make install
   if [ -e ./libexpect5.45.so ]; then
     cp libexpect5.45.so /usr/lib || exit 
   fi
@@ -34,48 +32,60 @@ installexpect()
   popd 
   rm -fr expect5.45*
 }
+
 installsshcopyid()
 {
+  echo "Installing ssh-copy-id if needed."
+  set +e
   hash ssh-copy-id > /dev/null 2>&1
   if [ $? -ne 0 ]; then
+    set -e
+    echo "ssh-copy-id not found - installing."
     cd /usr/bin
-    curl -o ssh-copy-id http://appscale.cs.ucsb.edu/appscale_files/ssh-copy-id || exit 1
+    curl -o ssh-copy-id http://appscale.cs.ucsb.edu/appscale_files/ssh-copy-id
     chmod +x ./ssh-copy-id
   fi
+  set -e
 }
+
 installsetuptools()
 {
-   mkdir -pv ${APPSCALE_HOME}/downloads
-   cd ${APPSCALE_HOME}/downloads
-   curl -o setuptools-0.6c11.tar.gz http://appscale.cs.ucsb.edu/appscale_files/setuptools-0.6c11.tar.gz  ||
- ext 1
-   tar zxvf setuptools-0.6c11.tar.gz
-   pushd setuptools-0.6c11
-   python setup.py install  || exit 1
-   popd
-   rm -fr  setuptools-0.6c11*
-}
-postinstallsetuptools()
-{
-  :;
+   echo "Installing setuptools if needed."
+   set +e
+   hash easy_install > /dev/null 2>&1
+   if [ $? -ne 0 ]; then
+     set -e
+     echo "setuptools not found - installing."
+     mkdir -pv ${APPSCALE_TOOLS_HOME}/downloads
+     cd ${APPSCALE_TOOLS_HOME}/downloads
+     curl -o setuptools-0.6c11.tar.gz http://appscale.cs.ucsb.edu/appscale_files/setuptools-0.6c11.tar.gz
+     tar zxvf setuptools-0.6c11.tar.gz
+     pushd setuptools-0.6c11
+     python setup.py install
+     popd
+     rm -fr setuptools-0.6c11*
+   fi
+   set -e
 }
 
-# only for the jaunty
 installec2ools()
 {
-# EC2
+  echo "Installing EC2 tools if needed."
+  set +e
   hash ec2-describe-instances > /dev/null 2>&1
-if [ $? -ne 0 ] && [ ! -f ${DESTDIR}/usr/local/bin/ec2-run-instances ]; then
-    mkdir -p ${APPSCALE_HOME}/downloads
-    cd ${APPSCALE_HOME}/downloads
+  if [ $? -ne 0 ] && [ ! -f ${DESTDIR}/usr/local/bin/ec2-run-instances ]; then
+    set -e
+    echo "EC2 tools not found - installing."
+    mkdir -p ${APPSCALE_TOOLS_HOME}/downloads
+    cd ${APPSCALE_TOOLS_HOME}/downloads
 
-    curl -o ec2-api-tools.zip http://s3.amazonaws.com/ec2-downloads/ec2-api-tools.zip || exit 1
-    curl -o ec2-ami-tools.zip http://s3.amazonaws.com/ec2-downloads/ec2-ami-tools.zip || exit 1
+    curl -o ec2-api-tools.zip http://s3.amazonaws.com/ec2-downloads/ec2-api-tools.zip
+    curl -o ec2-ami-tools.zip http://s3.amazonaws.com/ec2-downloads/ec2-ami-tools.zip
 
-    unzip ${APPSCALE_HOME}/downloads/ec2-api-tools*.zip || exit 1
-    unzip ${APPSCALE_HOME}/downloads/ec2-ami-tools*.zip || exit 1
-    rm -rf ${APPSCALE_HOME}/downloads/ec2-api-tools*.zip
-    rm -rf ${APPSCALE_HOME}/downloads/ec2-ami-tools*.zip
+    unzip ${APPSCALE_TOOLS_HOME}/downloads/ec2-api-tools*.zip
+    unzip ${APPSCALE_TOOLS_HOME}/downloads/ec2-ami-tools*.zip
+    rm -rf ${APPSCALE_TOOLS_HOME}/downloads/ec2-api-tools*.zip
+    rm -rf ${APPSCALE_TOOLS_HOME}/downloads/ec2-ami-tools*.zip
 
     mkdir -p ${DESTDIR}/usr/local/bin
     mkdir -p  ${DESTDIR}/usr/local/ec2-api-tools
@@ -84,11 +94,11 @@ if [ $? -ne 0 ] && [ ! -f ${DESTDIR}/usr/local/bin/ec2-run-instances ]; then
     rm -fr ${DESTDIR}/usr/local/ec2-ami-tools/*
     rm -fr ${DESTDIR}/usr/local/ec2-api-tools/*
 
-    mv -f ${APPSCALE_HOME}/downloads/ec2-ami-tools*/* ${DESTDIR}/usr/local/ec2-ami-tools || exit 1
-    mv -f ${APPSCALE_HOME}/downloads/ec2-api-tools*/* ${DESTDIR}/usr/local/ec2-api-tools || exit 1
+    mv -f ${APPSCALE_TOOLS_HOME}/downloads/ec2-ami-tools*/* ${DESTDIR}/usr/local/ec2-ami-tools
+    mv -f ${APPSCALE_TOOLS_HOME}/downloads/ec2-api-tools*/* ${DESTDIR}/usr/local/ec2-api-tools
 
-    rm -fr  ${APPSCALE_HOME}/downloads/ec2-ami-tools*/ || exit 1
-    rm -fr  ${APPSCALE_HOME}/downloads/ec2-api-tools*/ || exit 1    
+    rm -fr  ${APPSCALE_TOOLS_HOME}/downloads/ec2-ami-tools*/
+    rm -fr  ${APPSCALE_TOOLS_HOME}/downloads/ec2-api-tools*/
 
     mkdir -p ${DESTDIR}/etc/profile.d
     cat > ${DESTDIR}/etc/profile.d/ec2.sh <<EOF
@@ -99,72 +109,50 @@ EOF
     cp /usr/local/ec2-api-tools/bin/* ${DESTDIR}/usr/local/bin 
     cp /usr/local/ec2-ami-tools/bin/* ${DESTDIR}/usr/local/bin 
  fi
- easy_install boto==2.6
+ set -e
 }
 
 installpylibs()
 {
   easy_install termcolor
   easy_install paramiko
+  easy_install boto==2.6
 }
 
-installeuca2ools()
+installgem()
 {
-# Eucalyptus
-  hash euca-run-instances > /dev/null 2>&1
-  if [ $? -ne 0 ] && [ ! -f ${DESTDIR}/usr/local/bin/euca-run-instances ]; then
-    easy_install -U boto || exit 1
-    VERSION="1.3.1"
-    cd ${APPSCALE_HOME}/downloads
-    
-    # Install deps 
-    curl -o euca2ools-${VERSION}-src-deps.tar.gz http://appscale.cs.ucsb.edu/appscale_files/euca2ools-${VERSION}-src-deps.tar.gz || exit 1
-    tar zvxf euca2ools-$VERSION-src-deps.tar.gz
-    cd euca2ools-$VERSION-src-deps
-    tar zxvf boto-1.9b.tar.gz
-    cd boto-1.9b
-    python setup.py install
-    cd ..
-    tar zxvf M2Crypto-0.20.2.tar.gz
-    cd M2Crypto-0.20.2
-    python setup.py install
-    cd ..
-
-    curl -o euca2ools-${VERSION}.tar.gz http://appscale.cs.ucsb.edu/appscale_files/euca2ools-${VERSION}.tar.gz  || exit 1
-    tar zxvf euca2ools-${VERSION}.tar.gz || exit 1
-    rm -rf euca2ools-${VERSION}.tar.gz
-    cd euca2ools-${VERSION}
-    make PREFIX=${DESTDIR}/usr/local
-    easy_install euca2ools || exit 1
-    
-    cd ${APPSCALE_HOME}/downloads
-    rm -fr euca2ools-${VERSION}
+  echo "Installing gem if needed."
+  set +e
+  hash gem > /dev/null 2>&1
+  if [ $? -ne 0 ]; then
+    set -e
+    echo "gem not found - installing."
+    cd
+    wget http://appscale.cs.ucsb.edu/appscale_files/rubygems-1.3.7.tgz
+    tar zxvf rubygems-1.3.7.tgz
+    cd rubygems-1.3.7
+    ruby setup.rb
+    cd
+    ln -sf /usr/bin/gem1.8 /usr/bin/gem
+    rm -rf rubygems-1.3.7.tgz
+    rm -rf rubygems-1.3.7
   fi
+  set -e
 }
 
-postinstalleuca2ools()
+installrubylibs()
 {
-#    cd ${APPSCALE_HOME}/euca2ools-1.0-src-deps/boto-1.8d
-#    python setup.py install
-#    cd ${APPSCALE_HOME}/euca2ools-1.0
-#    make
-# just enable eggs
-    #easy_install boto
-    #easy_install euca2ools
-    :;
+  echo "Installing specific Ruby gems."
+  GEMDEST=${DESTDIR}/var/lib/gems/1.8
+  GEMOPT="--no-rdoc --no-ri"
+  gem install json flexmock ${GEMOPT}
+
+  # Rake 10.0 depecates rake/rdoctask - upgrade later
+  gem install -v=0.9.2.2 rake ${GEMOPT}
 }
 
 installappscaletools()
 {
-#    mkdir -p ${DESTDIR}/usr/local
-#    cd ${DESTDIR}/usr/local
-# tar ball is old right now.
-#    wget http://kings.cs.ucsb.edu/appscale_files/appscale-tools-1.3.tar.gz || exit 1
-#    rm appscale-tools-1.3.tar.gz
-# tools is copied by debian/rule file now.
-#    bzr branch lp:appscale/trunk-tools appscale-tools || exit 1
-#    rm -r appscale-tools/.bzr || exit 1
-
     # add to path
     mkdir -p ${DESTDIR}/etc/profile.d
     cat > ${DESTDIR}/etc/profile.d/appscale-tools.sh <<EOF
@@ -173,13 +161,8 @@ export PATH=\${PATH}:\${TOOLS_PATH}/bin
 EOF
 }
 
-postinstallappscaletools()
-{
-    :;
-}
-
 keygen()
 {
-    # create ssh private key if it is not exists.
+    # create ssh private key if it does not exist
     test -e /root/.ssh/id_rsa || ssh-keygen -q -t rsa -f /root/.ssh/id_rsa -N ""
 }
